requirements.txt
docker
watchdog
python-dotenv